Fix Bug

+++ src/Core/Route.php
@@ -152,7 +152,7 @@ namespace Npf\Core {
                     $this->appPath = explode("\\", $appFile);
                     break;
                 }
-                $lowerCase = substr($lowerCase, 0, strrpos($lowerCase, "\\", -2)) . "\\";
+                $lowerCase = substr($lowerCase, 0, (int)strrpos($lowerCase, "\\", -2)) . "\\";
             } while ($lowerCase != '\\');
         }
